requirements.txt
pip~=21.1.2
wheel~=0.36.2
setuptools~=57.0.0
Flask~=2.1.3
Pillow~=9.2.0
tensorflow~=2.9.1
keras~=2.9.0